+++ example/lib/pages/mini_games/lpc/lpc_game.dart
@@ -26,7 +26,6 @@ class LPCGame extends StatelessWidget {
   Widget build(BuildContext context) {
     return BonfireWidget(
       joystick: Joystick(
-        keyboardConfig: KeyboardConfig(),
         directional: JoystickDirectional(),
       ),
       map: WorldMapByTiled(