+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-# next
-- Export `package:flame/parallax.dart`
-- background improvements.
+# 3.5.0
+- Adds Parallax support
+- Background improvements.
+- [BREAKING CHANGE] move `keyboardConfig` to outside the `Joystick`. now in `BonfireWidget`.
+- [BREAKING CHANGE] moenamee `JoystickController` to `PlayerController`.
 
 # 3.4.0
 - Adds `header` param in `TiledNetworkReader`